# openapi.json
{"openapi":"3.1.0","info":{"title":"API Cadastro de Ninjas","description":"API REST para cadastro de ninjas e missões - Temática Naruto","contact":{"name":"Nadson Aguiar","url":"https://github.com/NadsonAguiar","email":"nadsonaguiar1372@gmail.com"},"license":{"name":"MIT License","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Ninjas","description":"Gerenciamento de ninjas"}],"paths":{"/ninjas/{id}":{"get":{"tags":["Ninjas"],"summary":"Lista um ninja por ID","description":"Rota lista um ninja por ID","operationId":"buscar","parameters":[{"name":"id","in":"path","description":"Usuário manda o id no caminho da requisição ","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Ninja encontrado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}},"404":{"description":"Ninja não encontrado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}}}},"put":{"tags":["Ninjas"],"summary":"Atualiza um ninja pelo seu ID","description":"Rota atualiza um ninja pelo seu ID","operationId":"atualizar","parameters":[{"name":"id","in":"path","description":"Usuário manda o id no caminho da requisição","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}},"required":true},"responses":{"201":{"description":"Ninja alterado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}},"400":{"description":"Ninja não encontrado, não foi possível alterar","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}}}},"delete":{"tags":["Ninjas"],"summary":"Deleta um ninja pelo seu ID","description":"Rota deleta um ninja pelo seu ID","operationId":"deletar","parameters":[{"name":"id","in":"path","description":"Usuário manda o ID no caminho da requisição","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Ninja deletado com sucesso"},"404":{"description":"Ninja não encontrado, não foi possível deletar"}}}},"/missoes/{id}":{"get":{"tags":["missoes-controller"],"operationId":"buscar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MissoesDTO"}}}}}},"put":{"tags":["missoes-controller"],"operationId":"atualizar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissoesDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MissoesDTO"}}}}}},"delete":{"tags":["missoes-controller"],"operationId":"deletar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ninjas":{"get":{"tags":["Ninjas"],"summary":"Lista todos os ninjas","description":"Rota lista todos os ninjas contidos no banco de dados","operationId":"listar","responses":{"200":{"description":"Ninjas listados com sucesso","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NinjaDTO"}}}}}}},"post":{"tags":["Ninjas"],"summary":"Cria um ninja","description":"Rota cria um ninja e inseri no banco de dados","operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}},"required":true},"responses":{"201":{"description":"Ninja criado com sucesso","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}},"400":{"description":"Erro na criação do ninja","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NinjaDTO"}}}}}}},"/missoes":{"get":{"tags":["missoes-controller"],"operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MissoesDTO"}}}}}}},"post":{"tags":["missoes-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissoesDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MissoesDTO"}}}}}}}},"components":{"schemas":{"MissoesModel":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"dificuldade":{"type":"string"},"rank":{"type":"string"}}},"NinjaDTO":{"type":"object","description":"Usuário cria o ninja usando o corpo da requisição","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string","description":"Nome do ninja","example":"Naruto Uzumaki","minLength":1},"email":{"type":"string","format":"email","description":"Email do ninja","example":"naruto@teste.com","minLength":1},"idade":{"type":"integer","format":"int32","description":"Idade do ninja","example":15},"rank":{"type":"string","description":"Rank do ninja","example":"Genin","minLength":1},"missoes":{"$ref":"#/components/schemas/MissoesModel"}},"required":["email","idade","nome","rank"]},"MissoesDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string","minLength":1},"dificuldade":{"type":"string","minLength":1},"rank":{"type":"string","minLength":1}},"required":["dificuldade","nome","rank"]}}}}